from_json for errors Document
`Alembic.FromJson.from_json/2` is a callback that all the modules under
`Alembic` can implement when using
`@behaviour Alembic.FromJson` to convert raw, decoded JSON to the module's
respective type, either a struct or map.

This commit does not fully implement the JSON API document, it only
implements enough to parse an errors document as from_json needs errors
documents to return when there is an error during conversion and
validation.